# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Veritas&Co Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professionals/me:
    get:
      operationId: getMyProfessionalProfile
      tags: [professionals]
      summary: Get current professional's profile
      responses:
        "200":
          description: Professional profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /professionals/passport/{passportId}:
    get:
      operationId: getPassport
      tags: [professionals]
      summary: Get public trust passport by ID
      parameters:
        - name: passportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trust passport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustPassport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /professionals:
    get:
      operationId: listProfessionals
      tags: [professionals]
      summary: Browse verified professionals (for clients)
      parameters:
        - name: skillsCategory
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
        - name: minRate
          in: query
          schema:
            type: integer
        - name: maxRate
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of professionals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionalListResponse"

  # --- VERIFICATION ---
  /verification/submit:
    post:
      operationId: submitVerification
      tags: [professionals]
      summary: Submit verification application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List open projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: preferredTier
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Post a new project (client only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # --- APPLICATIONS ---
...
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Blocked content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /messages/conversation/{userId}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get full conversation with a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"

  # --- WAITLIST ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Professional:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bio:
          type: ["string", "null"]
        skillsCategory:
          type: ["string", "null"]
        proficiencyLevel:
          type: ["string", "null"]
        portfolioLinks:
          type: array
          items:
            type: string
        workHistory:
          type: ["string", "null"]
        verificationStatus:
          type: string
          enum: [pending, verified, rejected, not_submitted]
        trustScore:
          type: ["integer", "null"]
        tier:
          type: ["string", "null"]
        projectsCompleted:
          type: integer
        deliveryRate:
          type: ["number", "null"]
        clientSatisfaction:
          type: ["number", "null"]
        passportId:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
      required: [id, userId, verificationStatus, projectsCompleted, name, email]
    TrustPassport:
      type: object
      properties:
        passportId:
          type: string
        displayName:
          type: string
        userId:
          type: integer
        country:
          type: ["string", "null"]
        trustScore:
          type: ["integer", "null"]
        tier:
          type: ["string", "null"]
        skillsCategory:
          type: ["string", "null"]
        projectsCompleted:
          type: integer
        deliveryRate:
          type: ["number", "null"]
        clientSatisfaction:
          type: ["number", "null"]
        identityVerified:
          type: boolean
        skillsAssessed:
          type: boolean
        backgroundCleared:
          type: boolean
        integrityConfirmed:
          type: boolean
      required: [passportId, displayName, projectsCompleted, identityVerified, skillsAssessed, backgroundCleared, integrityConfirmed]
    ProfessionalListResponse:
      type: object
      properties:
        professionals:
          type: array
          items:
            $ref: "#/components/schemas/Professional"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [professionals, total, page, limit]
    VerificationInput:
      type: object
      properties:
        bio:
          type: string
        country:
          type: string
        skillsCategory:
          type: string
        proficiencyLevel:
          type: string
        portfolioLinks:
          type: array
          items:
            type: string
        workHistory:
          type: string
      required: [bio, country, skillsCategory, proficiencyLevel, workHistory]
    Project:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: string
        skillsRequired:
          type: array
          items:
            type: string
        duration:
          type: string
        budgetMin:
          type: integer
        budgetMax:
          type: integer
        preferredTier:
          type: string
        status:
          type: string
          enum: [open, matched, active, completed]
        createdAt:
          type: string
      required: [id, clientId, title, description, duration, budgetMin, budgetMax, preferredTier, status, createdAt]
    ProjectInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        skillsRequired:
          type: array
          items:
            type: string
        duration:
          type: string
        budgetMin:
          type: integer
        budgetMax:
          type: integer
        preferredTier:
          type: string
      required: [title, description, duration, budgetMin, budgetMax, preferredTier]
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
      required: [projects, total, page]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, senderId, receiverId, content, isRead, createdAt]
    SendMessageInput:
      type: object
      properties:
        receiverId:
          type: integer
        content:
          type: string
      required: [receiverId, content]
...
    ConversationResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        otherUser:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required: [id, name]
      required: [messages, otherUser]
...