# lib/api-zod/src/generated/types/conversationResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Veritas&Co Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationResponseOtherUser } from './conversationResponseOtherUser';
import type { Message } from './message';

export interface ConversationResponse {
  messages: Message[];
  otherUser: ConversationResponseOtherUser;
}