# lib/api-zod/src/generated/types/verificationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Veritas&Co Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationInput {
  bio: string;
  country: string;
  skillsCategory: string;
  proficiencyLevel: string;
  portfolioLinks?: string[];
  workHistory: string;
}